--- a/src/libraries/Common/src/Interop/Unix/System.Native/Interop.GetOSArchitecture.cs
+++ b/src/libraries/Common/src/Interop/Unix/System.Native/Interop.GetOSArchitecture.cs
@@ -8,6 +8,7 @@ internal static partial class Interop
 {
     internal static partial class Sys
     {
+        [SuppressGCTransition]
         [DllImport(Libraries.SystemNative, EntryPoint = "SystemNative_GetOSArchitecture")]
         internal static extern int GetOSArchitecture();
     }