--- a/src/libraries/System.Runtime.InteropServices.RuntimeInformation/src/System/Runtime/InteropServices/RuntimeInformation/RuntimeInformation.Unix.cs
+++ b/src/libraries/System.Runtime.InteropServices.RuntimeInformation/src/System/Runtime/InteropServices/RuntimeInformation/RuntimeInformation.Unix.cs
@@ -11,8 +11,8 @@ namespace System.Runtime.InteropServices
         private static readonly object s_processLock = new object();
         private static string? s_osPlatformName;
         private static string? s_osDescription;
-        private static int s_osArch = -1;
-        private static int s_processArch = -1;
+        private static volatile int s_osArch = -1;
+        private static volatile int s_processArch = -1;
 
         public static bool IsOSPlatform(OSPlatform osPlatform)
         {