--- a/src/libraries/System.Runtime.InteropServices.RuntimeInformation/src/System/Runtime/InteropServices/RuntimeInformation/RuntimeInformation.Unix.cs
+++ b/src/libraries/System.Runtime.InteropServices.RuntimeInformation/src/System/Runtime/InteropServices/RuntimeInformation/RuntimeInformation.Unix.cs
@@ -7,8 +7,7 @@ namespace System.Runtime.InteropServices
 {
     public static partial class RuntimeInformation
     {
-        private static readonly object s_osLock = new object();
-        private static readonly object s_processLock = new object();
+        private static readonly object s_lock = new object();
         private static string? s_osPlatformName;
         private static string? s_osDescription;
         private static volatile int s_osArch = -1;
@@ -30,7 +29,7 @@ namespace System.Runtime.InteropServices
 
                 if (s_osArch == -1)
                 {
-                    lock (s_osLock)
+                    lock (s_lock)
                     {
                         if (s_osArch == -1)
                         {
@@ -70,7 +69,7 @@ namespace System.Runtime.InteropServices
 
                 if (s_processArch == -1)
                 {
-                    lock (s_processLock)
+                    lock (s_lock)
                     {
                         if (s_processArch == -1)
                         {