```cs
[Benchmark]
public Architecture OSArchitecture() => RuntimeInformation.OSArchitecture;

[Benchmark]
public Architecture ProcessArchitecture() => RuntimeInformation.ProcessArchitecture;
```

BEFORE:

|              Method |      Mean |     Error |    StdDev |    Median |       Min |       Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |----------:|----------:|----------:|----------:|----------:|----------:|------:|------:|------:|----------:|
|      OSArchitecture | 16.202 ns | 0.1181 ns | 0.1105 ns | 16.184 ns | 16.073 ns | 16.362 ns |     - |     - |     - |         - |
| ProcessArchitecture | 16.221 ns | 0.0320 ns | 0.0267 ns | 16.217 ns | 16.180 ns | 16.288 ns |     - |     - |     - |         - |

AFTER:

|              Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |---------:|----------:|----------:|---------:|---------:|---------:|------:|------:|------:|----------:|
|      OSArchitecture | 1.059 ns | 0.0073 ns | 0.0065 ns | 1.059 ns | 1.049 ns | 1.070 ns |     - |     - |     - |         - |
| ProcessArchitecture | 1.331 ns | 0.0129 ns | 0.0120 ns | 1.328 ns | 1.314 ns | 1.354 ns |     - |     - |     - |         - |